fix comment: transfer TxnBatchedPositionImpl to outer public class
