+++ pulsar-transaction/coordinator/src/main/java/org/apache/pulsar/transaction/coordinator/impl/TxnLogBufferedWriter.java
@@ -421,40 +421,7 @@ public class TxnLogBufferedWriter<T> implements AsyncCallbacks.AddEntryCallback,
         }
     }
 
-    /***
-     * The difference with {@link PositionImpl} is that there are two more parameters:
-     * {@link #batchSize}, {@link #batchIndex}.
-     */
-    public static class TxnBatchedPositionImpl extends PositionImpl {
 
-        /** The data length of current batch. **/
-        @Getter
-        private final int batchSize;
-
-        /** The position of current batch. **/
-        @Getter
-        private final int batchIndex;
-
-        public TxnBatchedPositionImpl(Position position, int batchSize, int batchIndex, long[] ackSet){
-            super(position.getLedgerId(), position.getEntryId(), ackSet);
-            this.batchIndex = batchIndex;
-            this.batchSize = batchSize;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (o instanceof TxnBatchedPositionImpl other) {
-                return super.equals(o) && batchSize == other.batchSize && batchIndex == other.batchIndex;
-            }
-            return false;
-
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(super.hashCode(), batchSize, batchIndex);
-        }
-    }
 
     interface AddDataCallback {
