+++ pulsar-transaction/coordinator/src/test/java/org/apache/pulsar/transaction/coordinator/impl/TxnLogBufferedWriterTest.java
@@ -155,8 +155,8 @@ public class TxnLogBufferedWriterTest extends MockedBookKeeperTestCase {
             ArrayList<Position> innerPositions = callbackPositionIterator.next();
             int batchSize = entryDataArray.length;
             for(int i = 0; i < entryDataArray.length; i++){
-                TxnLogBufferedWriter.TxnBatchedPositionImpl innerPosition =
-                        (TxnLogBufferedWriter.TxnBatchedPositionImpl) innerPositions.get(i);
+                TxnBatchedPositionImpl innerPosition =
+                        (TxnBatchedPositionImpl) innerPositions.get(i);
                 Assert.assertEquals(innerPosition.getBatchSize(), batchSize);
                 Assert.assertEquals(innerPosition.getBatchIndex(), i);
             }
